refactor: rename vercelFormatter to vercel-formatter (kebab-case)
