--- batteries/vercelFormatter/package.json
@@ -1,5 +1,0 @@
-{
-	"types": "../../dist/esm/batteries/vercelFormatter.d.ts",
-	"main": "../../dist/cjs/batteries/vercelFormatter.js",
-	"module": "../../dist/esm/batteries/vercelFormatter.js"
-}
