+++ example/vercel-formatter.ts
@@ -1,4 +1,4 @@
-import { vercelFormatter } from "../batteries/vercelFormatter";
+import { vercelFormatter } from "../batteries/vercel-formatter";
 import { setDefaultHelpFormatter } from "../src";
 
 setDefaultHelpFormatter(vercelFormatter);
